+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## 5.0.0+1
+## 5.0.0+2
 September 04, 2024
 - Remove all deprecated properties, functions, and methods
 - Introduce RouteObserverStates.subscribeRoutes(this); in route_observer_states.dart
@@ -14,6 +14,10 @@ September 04, 2024
 - Rename bool forEach() to bool forEachState()
 - mixin RecordExceptionMixin to mixin RecordExceptionMixin on State
 - Update README Example app
+- Separated tests into some Dart files
+- Renamed part files
+- Commented out mixin StateXRouteAware 
+- Commented out void '_copyOver' functions
 
 ## 4.24.0
 August 18, 2024